+++ include/strings.h
@@ -3227,4 +3227,5 @@ extern const u8 gText_MoveItemWhere[];
 extern const u8 gText_XsYAnd[];
 extern const u8 gText_XsYWereSwapped[];
 
+extern const u8 gText_Relearn_Menu[];
 #endif // GUARD_STRINGS_H