+++ src/move_relearner.c
@@ -25,6 +25,7 @@
 #include "constants/rgb.h"
 #include "constants/songs.h"
 #include "daycare.h"
+#include "party_menu.h"
 
 /*
  * Move relearner state machine
@@ -680,6 +681,7 @@ static void DoMoveRelearnerMain(void)
         if (!gPaletteFade.active)
         {
             FreeMoveRelearnerResources();
+            if (FlagGet(FLAG_TEMP_1)) { CB2_ReturnToPartyMenuFromSummaryScreen(); break;}
             SetMainCallback2(CB2_ReturnToField);
         }
         break;